+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - fix bug with commit --force when tests throws exception
 - fix bug - when you import authored component (and there is newer version) it duplicate it in the .bit.map.json
 - fix bug - when you import authored component it was added to bit.json dependencies
+- create package.json for imported components
 
 ## [0.10.7] - 2017-09-07